config: migrate digest class for ActiveSupport::Digest changing to SHA256
This change affects the encrypted cookies and etags.
If our application uses it, we should add the rotator from SHA1 to SHA256.

In Ranguba, we don't use cookies and etags which are related to this
change. So I think we don't have to deal with it.

If we have to add the rorater for the Ranguba users who add the
additonal functions using ActiveSupport::Digest, please let me know.

ref: https://guides.rubyonrails.org/v7.0/upgrading_ruby_on_rails.html#key-generator-digest-class-changing-to-use-sha256
ref: https://guides.rubyonrails.org/v7.0/upgrading_ruby_on_rails.html#digest-class-for-activesupport-digest-changing-to-sha256