config action controller: enable parameter wrapping for JSON by default
This setting is being changed to default in Rails v7.
Setting this configuration value to true has the same behavior as the
initializer. So `config/initializers/wrap_parameters.rb` will not be
generated anymore in Rails v7.

There is no effect to Ranguba because we don't use any custom parameter
wrappings.

ref: https://guides.rubyonrails.org/configuring.html#config-action-controller-wrap-parameters-by-default