config action dispatch: migrate cookie serializer from `:marshal` to `:json`
GitHub: ref GH-7

In Rails v7, this setting defaults to `:json`.

Using `:marshal` as the default value was for backward compatibility with Rails v4.0.
We no longer need this default because we have been using the JSON format by overwriting it in `config/initializers/cookies_serializer.rb`.

In Ranguba, this change has no effect since we don't use cookies.

ref: https://github.com/rails/rails/pull/42538
ref: https://guides.rubyonrails.org/configuring.html#config-action-dispatch-cookies-serializer