config active support: disables the deprecated #to_s override
GitHub: ref GH-7

In Rails v7, this setting is default.

Disables the override of the #to_s methods in some Ruby core classes.
This config is for applications that want to take advantage early of a Ruby 3.1 optimization.
- ref: https://github.com/rails/rails/pull/43772

In Ranguba, there is no effect because we don't use the overrided method
function like passing `format` to `to_s`.

ref: https://guides.rubyonrails.org/v7.0.2/configuring.html#config-active-support-disable-to-s-conversion