config application: remove autoload paths from $LOAD_PATH (#55)
GitHub: ref GH-34

In Rails v7.1, this setting is default.
ref:
https://guides.rubyonrails.org/v7.1/configuring.html#config-add-autoload-paths-to-load-path

Zeitwerk uses absolute paths internally, and applications running in
`:zeitwerk` mode do not need require_dependency, so models, controllers,
jobs, etc. do not need to be in `$LOAD_PATH`.
That's why we don't have to add autoload paths to `$LOAD_PATH` anymore
using Zeitwerk.

Note: The `lib` directory is not affected by this flag, it is added to
`$LOAD_PATH` always.